Alteracao tipo do atributo peso da classe produto
Alteracao do tipo do atributo peso da classe produto, junto com a inclusao de mais uma linha de comentario explicando a classe

+++ PastaFontesApp/Produto.java
@@ -1,20 +1,21 @@
 //Classe que guarda dados de produtos e serviços oferecidos no estacionamento
+//Exemplos de produtos e serviços: estacionar veiculo; lavar veiculo; trocar oleo
 
 public class Produto {
 
-	private double peso;
+	private int peso;
 	private double altura;
 	
 	public Produto() {
 		
 	}
 	
-	public Produto(double peso, double altura) {
+	public Produto(int peso, double altura) {
 		this.peso=peso;
 		this.altura=altura;
 	}
 	
-	public double getPeso() {
+	public int getPeso() {
 		return peso;
 	}
